assistants/km_agno/pyproject.toml
[project]
name = "km-agno"
version = "0.1.0"
description = "Index flink-studies for semantic search with Agno and Chroma"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno[chromadb,openai,readers]>=2.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.3.0",
    "pytest>=9.0.3",
    "pypdf>=6.10.2",
    "unstructured>=0.22.26",
    "markdown>=3.10.2",
    "sqlalchemy>=2.0.49",
]

[project.scripts]
km_agno = "km_agno.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["km_agno"]

# Ensure `import km_agno` works for pytest and the IDE when the project root is on sys.path
# (matches editable install layout without requiring `uv sync` for discovery).
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.basedpyright]
extraPaths = ["."]

